+++ Resources/doc/changes.md
@@ -1,6 +1,24 @@
 CHANGE LOG
 ==========
 
+V 1.3.0
+-------
+
+Attention! Breaking Changes (BC) below!
+
+- Changed getPageIndex()
+  with $key default to 'pageindex'
+- Changed resetPageIndex()
+  with $key default to 'pageindex'
+- BC: Changed getPageRange()
+  added $key = 'pagerange'
+  now using auto_register
+- BC: Changed getPageSize()
+  added $key = 'pagesize'
+  now using auto_register
+- Added getPageSizeSelector()
+  Method to output the page-size-selector html
+
 V 1.2.0
 -------